requirements.txt
streamlit
python-docx
pdf2image
pytesseract
openai
spacy
opencv-python
numpy
